Create 1949-Strong Friendship
Notice the differences with 1951.

Problem 1951 is users and their followers, which is the one-way connection. But this is a two-way friendship.

Therefore, this problem needs to use the 'union' first and repeat the process of finding common followers. But remember to limit the combination of U1 and U2 is appeared in the union dataset.